+++ Resources/Setup/Terraform-Azure/script3_sql_server.tf
@@ -11,8 +11,8 @@ resource "azurerm_mssql_server" "sqlsrv" {
     object_id      = data.azuread_service_principal.sp_admin.object_id
     #login_username = data.azuread_user.admin_user.user_principal_name
     #object_id      = data.azuread_user.admin_user.object_id
-    #login_username              = azuread_group.g5.display_name
-    #object_id                   = azuread_group.g5.object_id
+    #login_username              = data.azuread_group.g5.display_name
+    #object_id                   = data.azuread_group.g5.object_id
     azuread_authentication_only = true
   }
   identity {
@@ -24,10 +24,18 @@ resource "azurerm_mssql_server" "sqlsrv" {
 
 resource "azurerm_role_assignment" "sqlsrv_admin" {
   role_definition_name = "Contributor"
-  #principal_id         = azuread_group.g5.id
+  #principal_id         = data.azuread_group.g5.id
   principal_id         = data.azuread_service_principal.sp_admin.id
   scope                = azurerm_mssql_server.sqlsrv.id
 }
+
+# Assign the Directory Readers role to the Managed Identity of the SQL Server
+resource "azurerm_role_assignment" "sqlsrv_directory_reader" {
+  role_definition_name = "Reader"
+  principal_id         = azurerm_mssql_server.sqlsrv.identity[0].principal_id  
+  scope                = azurerm_mssql_server.sqlsrv.id
+}
+
 
 resource "azurerm_mssql_firewall_rule" "sqlsrv_firewall_rule" {
   name             = "fw-${var.org}-${var.app}-sqlsrv-allow-${var.env}"
@@ -36,13 +44,20 @@ resource "azurerm_mssql_firewall_rule" "sqlsrv_firewall_rule" {
   end_ip_address   = data.external.external_ip.result.ip
 }
 
+# Allow all domain users to connect to the SQL server. Restrict later to the organization range.
+resource "azurerm_mssql_firewall_rule" "sqlsrv_firewall_rule_all" {
+  name             = "fw-${var.org}-${var.app}-sqlsrv-allow-all-${var.env}"
+  server_id        = azurerm_mssql_server.sqlsrv.id
+  start_ip_address = "0.0.0.0"
+  end_ip_address   = "255.255.255.255"
+}
+
 # Assign SQL Server roles inside SQL Server by using SQL script.
 
 resource "null_resource" "assign_sql_server_roles" {
   provisioner "local-exec" {
     command     = <<-EOT
-    $tokenResponse = az account get-access-token --resource https://database.windows.net/ --output json
-    $token = ($tokenResponse | ConvertFrom-Json).accessToken
+    $token = $env:ARM_DATABASE_ACCESS_TOKEN
     $serverName = "${azurerm_mssql_server.sqlsrv.fully_qualified_domain_name}"
     $sqlCommandText = Get-Content "script3_sql_server_roles.sql" -Raw
     Invoke-Sqlcmd -ServerInstance $serverName -Database 'master' -AccessToken $token -Query $sqlCommandText
@@ -50,6 +65,10 @@ resource "null_resource" "assign_sql_server_roles" {
     interpreter = ["PowerShell", "-Command"]
   }
   depends_on = [
-    azurerm_mssql_server.sqlsrv
+    azurerm_mssql_server.sqlsrv,
+    azurerm_role_assignment.sqlsrv_directory_reader
   ]
 }
+
+#$tokenResponse = az account get-access-token --resource https://database.windows.net/ --output json
+#$token = ($tokenResponse | ConvertFrom-Json).accessToken